Fix the stride issues Handle more cases on the core pool

+++ src/lib/renderer/enesim_renderer_raddist.c
@@ -46,7 +46,7 @@ static void _span_identity(Enesim_Renderer *r, int x, int y,
 	double r_inv;
 	uint32_t *src;
 	int sw, sh;
-	int sstride;
+	size_t sstride;
 	double ox, oy;
 
 	rd = _raddist_get(r);
